+++ tests/Header/ContentLanguageTest.php
@@ -6,7 +6,7 @@ namespace Tests\Innmind\Http\Header;
 use Innmind\Http\{
     Header\ContentLanguage,
     Header,
-    Header\ContentLanguageValue
+    Header\Content\Language,
 };
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
@@ -14,24 +14,19 @@ class ContentLanguageTest extends TestCase
 {
     public function testInterface()
     {
-        $h = new ContentLanguage(
-            $v = new ContentLanguageValue('fr'),
-        );
+        $h = Language::maybe('fr')
+            ->map(ContentLanguage::of(...))
+            ->match(
+                static fn($header) => $header,
+                static fn() => null,
+            );
 
         $this->assertInstanceOf(Header\Custom::class, $h);
         $this->assertSame('Content-Language: fr', $h->normalize()->toString());
     }
 
-    public function test()
-    {
-        $header = ContentLanguage::of('fr');
-
-        $this->assertInstanceOf(ContentLanguage::class, $header);
-        $this->assertSame('Content-Language: fr', $header->normalize()->toString());
-    }
-
     public function testWithoutValues()
     {
-        $this->assertSame('Content-Language: ', (new ContentLanguage)->normalize()->toString());
+        $this->assertSame('Content-Language: ', ContentLanguage::of()->normalize()->toString());
     }
 }